Move to gophers graphql fork and check for field
Implement check for field so items from receipts are selected only where
the field exists. For this I needed to create a fork which uses a
feature that is not yet pulled into master of gophers graphql. This is a
very hacky solution and it should be reset to gophers' module the moment
that feature hits the master branch.